+++ src/components/trip.js
@@ -1,0 +1,51 @@
+const getTripDaysHTML = () => `<ul class="trip-days"></ul>`;
+
+const getTripDayHTML = () => `<li class="trip-days__item  day">
+  <div class="day__info">
+    <span class="day__counter">1</span>
+    <time class="day__date" datetime="2019-03-18">MAR 18</time>
+  </div>
+  <ul class="trip-events__list"></ul>
+</li>`;
+
+const getTripInfoHTML = () => `<div class="trip-info__main">
+  <h1 class="trip-info__title">Amsterdam &mdash; ... &mdash; Amsterdam</h1>
+  <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;21</p>
+</div>`;
+
+const getTripCardHTML = () => `<li class="trip-events__item">
+  <div class="event">
+    <div class="event__type">
+      <img class="event__type-icon" width="42" height="42" src="img/icons/taxi.png" alt="Event type icon">
+    </div>
+    <h3 class="event__title">Taxi to airport</h3>
+
+    <div class="event__schedule">
+      <p class="event__time">
+        <time class="event__start-time" datetime="2019-03-18T10:30">10:30</time>
+        &mdash;
+        <time class="event__end-time" datetime="2019-03-18T11:00">11:00</time>
+      </p>
+      <p class="event__duration">1H 30M</p>
+    </div>
+
+    <p class="event__price">
+      &euro;&nbsp;<span class="event__price-value">20</span>
+    </p>
+
+    <h4 class="visually-hidden">Offers:</h4>
+    <ul class="event__selected-offers">
+      <li class="event__offer">
+        <span class="event__offer-title">Order Uber</span>
+        &plus;
+        &euro;&nbsp;<span class="event__offer-price">20</span>
+       </li>
+    </ul>
+
+    <button class="event__rollup-btn" type="button">
+      <span class="visually-hidden">Open event</span>
+    </button>
+  </div>
+</li>`;
+
+export {getTripCardHTML, getTripDayHTML, getTripDaysHTML, getTripInfoHTML};
