+++ src/components/utils.js
@@ -1,0 +1,3 @@
+const render = (html, target, where = `beforeend`) => target.insertAdjacentHTML(where, html);
+
+export {render};
